+++ CHANGELOG.md
@@ -7,14 +7,19 @@ Formatted as described on http://keepachangelog.com/.
 
 ### Added
 
+- Webservice endpoint to render 2D fragment in SVG
 - Published documentation on http://kripodb.readthedocs.io
 - Documented update pipeline
 - Documented command line interface in Sphinx
+- Retrieve fragments from webservice based on fragment id and pdb code (#35)
 
 ### Changed
 
 - merge similarity pairs files in chunks instead of loading whole source file in memory
+- canned fragments_by_* methods can use local file or webservice
 - error when duplicate fragment insert is performed
+- Renamed `kripodb similarities serve` to `kripodb serve`, as it now also serves the fragments
+- Switched from nosetest to py.test (#36)
 
 ### Removed